Refactor: Update weather coordinates to Double and improve catch caching
This commit introduces two main improvements:

*   **Refactor: Use Double for weather coordinates**
    The `WeatherForecast` data class now uses `Double` instead of `String` for latitude and longitude. This change simplifies the code and avoids unnecessary type conversions, as seen in the `GetNewCatchWeatherUseCase`.

*   **Fix: Gracefully handle catch caching failures**
    Catch insertion operations into the local Room database are now wrapped in `try-catch` blocks within the `SyncAwareCatchesRepository`. This prevents the app from crashing if a catch fails to be saved locally, for example, due to a missing foreign key when its associated marker is not yet cached. Errors are now logged without interrupting the user flow.

+++ app/src/main/java/com/mobileprism/fishing/domain/use_cases/catches/GetNewCatchWeatherUseCase.kt
@@ -93,8 +93,8 @@ class GetNewCatchWeatherUseCase(
         val weather = lastLoadedWeather ?: return true
 
         val lastLoadedWeatherPlace = UserMapMarker(
-            latitude = weather.latitude.toDouble(),
-            longitude = weather.longitude.toDouble()
+            latitude = weather.latitude,
+            longitude = weather.longitude
         )
         return place?.let { currentPlace ->
             isLocationsTooFar(currentPlace, lastLoadedWeatherPlace)
